Create flag macro
We want to have sticky notes on all of our abort procedures so you can
very quickly flip to them from a closed book, and they should be colour
coded so you know which to flip to. This macro creates a coloured square
(selectable colour) on the right edge of the page to make it easier to
know where to put what colour sticky notes. The square is over the right
edge, so it should be visible even when book is closed, in case sticky
note falls off.

Only use this macro at the end of a line, or your formatting will be
screwed up.