Merge 2018 IREC procedures (#8)
* Add first draft 2018 IREC procedures

* Add checks procedure

* Clarify RLCS client box state during pre-launch checks

* Reorder valve turns during manual abort procedures

By opening the vent valve before the fill valve, the operators only turn
one pressurized valve instead of two. It isn't clear how much of a
safety win this is (probably extremely minor), but every little bit
counts

* Rework fill-time leaks check procedure

The best PPE is distance

Have the leak check be performed remotely, with primary and secondary
standing by at a safe distance (not safe for launch, but safe for a
leak) where they are still able to intervene to perform the abort
procedure. This change also required rework to abort procedure A4, since
the remote operator needs to safe the system before the operators can
approach

* Open tank vent before opening remote fill to fill

* Add tables

Latex tables are... not great

* Small tweaks to format and wording

* Add abort procedure for what to do when radio is lost

* Jacob review comments

- Fix missing end quotes
- Don't use the word we

* Small changes